pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
    "flask-bootstrap",
    "validators",
    "python-dateutil",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
